feat: save a node as good or bad
