+++ composeApp/src/commonMain/kotlin/proj/memorchess/axl/core/data/IStoredMove.kt
@@ -12,6 +12,12 @@ interface IStoredMove {
   /** The move in standard notation */
   val move: String
 
-  /** Whether the move has to be learned */
-  val isGood: Boolean
+  /**
+   * Whether the move has to be learned.
+   *
+   * A bad move is a mistake. It is still saved because the user has to learn how to counter it.
+   *
+   * Bad moves are always isolated: previous and the next moves are good.
+   */
+  var isGood: Boolean
 }
