+++ composeApp/src/commonMain/kotlin/proj/memorchess/axl/core/data/StoredMove.kt
@@ -1,10 +1,11 @@
 package proj.memorchess.axl.core.data
 
+/** Implementation of [IStoredMove] that can be used in core. */
 data class StoredMove(
   private val origin: PositionKey,
   private val destination: PositionKey,
   override val move: String,
-  override val isGood: Boolean = true,
+  override var isGood: Boolean = true,
 ) : IStoredMove {
   override fun getOrigin(): PositionKey {
     return origin
